+++ tests/Dom/DomEventTest.ts
@@ -1,11 +1,12 @@
 import {AbstractDomTestCase} from "../../dist";
+import {SpyInterface} from "../../src/Spy/SpyInterface";
 
 export class DomEventTest extends AbstractDomTestCase {
 
     public testCancelableFalseByDefault() {
         let doc = this.createDocumentWithCheckboxAndLabel();
         let event = new MouseEvent("click");
-        let visitor = new Visitor();
+        let visitor = this.createVisitor();
         this.initCheckboxAndLabelDispatchEvent(doc, event, visitor);
         this.assertTrue(visitor.isCalled());
     }
@@ -13,7 +14,7 @@ export class DomEventTest extends AbstractDomTestCase {
     public testSetCancelableTrue() {
         let doc = this.createDocumentWithCheckboxAndLabel();
         let event = new MouseEvent("click", {"cancelable": true});
-        let visitor = new Visitor();
+        let visitor = this.createVisitor();
         this.initCheckboxAndLabelDispatchEvent(doc, event, visitor);
         this.assertFalse(visitor.isCalled());
     }
@@ -21,7 +22,7 @@ export class DomEventTest extends AbstractDomTestCase {
     public testBubblesFalseByDefault() {
         let doc = this.createDocumentWithCheckboxAndLabel();
         let event = new MouseEvent("click");
-        let visitor = new Visitor();
+        let visitor = this.createVisitor();
         let label = doc.getElementById("label");
         label.addEventListener("click", () => {
             visitor.call();
@@ -34,7 +35,7 @@ export class DomEventTest extends AbstractDomTestCase {
     public testSetBubblesTrue() {
         let doc = this.createDocumentWithCheckboxAndLabel();
         let event = new MouseEvent("click", {"bubbles": true});
-        let visitor = new Visitor();
+        let visitor = this.createVisitor();
         let label = doc.getElementById("label");
         label.addEventListener("click", () => {
             visitor.call();
@@ -49,7 +50,7 @@ export class DomEventTest extends AbstractDomTestCase {
         this.assertInstanceOf(MouseEvent, event);
     }
 
-    private initCheckboxAndLabelDispatchEvent(doc: Document, event: Event, visitor: Visitor) {
+    private initCheckboxAndLabelDispatchEvent(doc: Document, event: Event, visitor: SpyInterface) {
         let checkbox = doc.getElementById("checkbox");
         let label = doc.getElementById("label");
         label.addEventListener("click", function (event) {
@@ -68,17 +69,5 @@ export class DomEventTest extends AbstractDomTestCase {
             <label for="checkbox" id="label">Click me<span id="nestedToLabelElement"></span></label>
         `);
         return docBuilder.getMock();
-    }
-}
-
-class Visitor {
-    private called: boolean = false;
-
-    public call(): void {
-        this.called = true;
-    }
-
-    public isCalled(): boolean {
-        return this.called;
     }
 }